+++ feeder/src/main/java/org/n52/sos/importer/feeder/Configuration.java
@@ -49,7 +49,11 @@ import javax.xml.namespace.QName;
 import org.apache.xmlbeans.XmlError;
 import org.apache.xmlbeans.XmlException;
 import org.apache.xmlbeans.XmlOptions;
-import org.n52.sos.importer.feeder.csv.WrappedCSVParser;
+import org.n52.sos.importer.feeder.collector.DefaultCsvCollector;
+import org.n52.sos.importer.feeder.collector.SampleBasedObservationCollector;
+import org.n52.sos.importer.feeder.collector.csv.WrappedCSVParser;
+import org.n52.sos.importer.feeder.importer.SingleObservationImporter;
+import org.n52.sos.importer.feeder.importer.SweArrayObservationWithSplitExtensionImporter;
 import org.n52.sos.importer.feeder.model.Offering;
 import org.n52.sos.importer.feeder.model.Position;
 import org.n52.sos.importer.feeder.model.Sensor;
