+++ feeder/src/main/java/org/n52/sos/importer/feeder/collector/DefaultCsvCollector.java
@@ -26,7 +26,7 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
  * Public License for more details.
  */
-package org.n52.sos.importer.feeder;
+package org.n52.sos.importer.feeder.collector;
 
 import java.io.IOException;
 import java.text.ParseException;
@@ -36,11 +36,13 @@ import java.util.Optional;
 import java.util.concurrent.CountDownLatch;
 
 import org.n52.oxf.om.x20.OmParameter;
+import org.n52.sos.importer.feeder.Application;
+import org.n52.sos.importer.feeder.Configuration;
+import org.n52.sos.importer.feeder.DataFile;
 import org.n52.sos.importer.feeder.model.FeatureOfInterest;
 import org.n52.sos.importer.feeder.model.InsertObservation;
 import org.n52.sos.importer.feeder.model.ObservedProperty;
 import org.n52.sos.importer.feeder.model.Offering;
-import org.n52.sos.importer.feeder.model.Position;
 import org.n52.sos.importer.feeder.model.Sensor;
 import org.n52.sos.importer.feeder.model.Timestamp;
 import org.n52.sos.importer.feeder.model.UnitOfMeasurement;
@@ -48,18 +50,13 @@ import org.n52.sos.importer.feeder.util.LineHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class SingleProfileCollector extends CollectorSkeleton {
+public class DefaultCsvCollector extends CollectorSkeleton {
 
-    private static final Logger LOG = LoggerFactory.getLogger(SingleProfileCollector.class);
-
-    private Timestamp timestamp;
-
-    private Sensor sensor;
-
-    private FeatureOfInterest foi;
+    static final Logger LOG = LoggerFactory.getLogger(DefaultCsvCollector.class);
 
     @Override
-    public void collectObservations(DataFile dataFile, CountDownLatch latch) throws IOException, ParseException {
+    public void collectObservations(DataFile dataFile, CountDownLatch latch)
+            throws IOException {
         if (configuration == null) {
             LOG.error("Configuration not set!");
             return;
@@ -73,13 +70,24 @@ public class SingleProfileCollector extends CollectorSkeleton {
             String[] headerLine = new String[0];
             this.dataFile = dataFile;
             lineCounter = dataFile.getFirstLineWithData();
+            if (dataFile.getHeaderLine() > -1) {
+                headerLine = readHeaderLine(dataFile);
+            }
             if (!dataFile.areMeasureValuesAvailable()) {
                 LOG.error("No measured value columns found in configuration");
                 return;
             }
+            int lastLine = context.getLastReadLine();
+            if (configuration.isUseLastTimestamp()) {
+                // pointing back on first line with data because we are skipping by date and not line
+                lastLine = dataFile.getFirstLineWithData();
+            }
+            if (configuration.getFirstLineWithData() == 0) {
+                skipLines(lastLine + 1);
+            } else {
+                skipLines(lastLine);
+            }
             String[] line;
-            processSampleStart();
-            skipLines(2);
             while ((line = parser.readNext()) != null && !stopped) {
                 if (!configuration.isLineIgnorable(line) &&
                         configuration.containsData(line) &&
@@ -100,80 +108,51 @@ public class SingleProfileCollector extends CollectorSkeleton {
                     LOG.info("Handled line {}.", lineCounter);
                 }
             }
+            context.setLastReadLine(lineCounter);
         } finally {
             latch.countDown();
         }
     }
 
-    private void processSampleStart() throws IOException {
-        String[] line;
-        // long, lat, alt
-        double[] values = new double[3];
-        while ((line = parser.readNext()) != null) {
-            if (line[0].equalsIgnoreCase(" Device")) {
-                LOG.debug("Creating sensor from '{}'", Arrays.toString(line));
-                sensor = new Sensor(line[1], line[1]);
-            } else if (line[0].equalsIgnoreCase(" Cast time (UTC)")) {
-                LOG.debug("Creating timestamp from '{}'", Arrays.toString(line));
-                timestamp = new Timestamp(line[1].replaceAll(" ", "T").concat("Z"));
-                timestamp.setTimezone(0);
-            } else if (line[0].equalsIgnoreCase(" Start latitude")) {
-                values[1] = Double.parseDouble(line[1]);
-            } else if (line[0].equalsIgnoreCase(" Start longitude")) {
-                values[0] = Double.parseDouble(line[1]);
-            } else if (line[0].equalsIgnoreCase(" Start altitude")) {
-                values[2] = Double.parseDouble(line[1]);
-            } else if (line[0].equalsIgnoreCase(" ")) {
-                break;
-            }
-        }
-        String deg = "deg";
-        Position position = new Position(values , new String[] {deg, deg, "m"}, 4326);
-        String identifier = generateFeatureIdentifier(values);
-        foi = new FeatureOfInterest(identifier, identifier, position);
-        foi.setParentFeature(configuration.getParentFeatureFromAdditionalMetadata());
-    }
-
-    private String generateFeatureIdentifier(double[] values) {
-        return new StringBuilder("profile-observation-at-")
-                .append(values[1])
-                .append("-")
-                .append(values[0])
-                .toString()
-                .replaceAll("\\.|,", "_");
-    }
-
     @Override
     protected InsertObservation getInsertObservationForMeasuredValue(int measureValueColumn, String[] line)
             throws ParseException {
+        LOG.trace("getInsertObservationForMeasuredValue(..)");
         // TIMESTAMP
-        if (configuration.isUseLastTimestamp() && !verifyTimeStamp(timestamp)) {
+        Timestamp timeStamp = dataFile.getTimeStamp(measureValueColumn, line);
+        if (configuration.isUseLastTimestamp() && !verifyTimeStamp(timeStamp)) {
             return null;
         }
-        LOG.debug("Timestamp: {}", timestamp);
+        // TODO implement using different templates in later version depending on the class of value
+        LOG.debug("Timestamp: {}", timeStamp);
+        // SENSOR
+        Sensor sensor = dataFile.getSensorForColumn(measureValueColumn, line);
         LOG.debug("Sensor: {}", sensor);
+        // FEATURE OF INTEREST incl. Position
+        FeatureOfInterest foi = dataFile.getFoiForColumn(measureValueColumn, line);
         LOG.debug("Feature of Interest: {}", foi);
         // VALUE
-        final Object value = dataFile.getValue(measureValueColumn, line);
+        Object value = dataFile.getValue(measureValueColumn, line);
         if (value.equals(Configuration.SOS_OBSERVATION_TYPE_NO_DATA_VALUE)) {
             return null;
         }
+        // TODO implement handling for value == null => skip observation and log it, or logging is done in getValue(..)
         LOG.debug("Value: {}", value.toString());
         // UOM CODE
-        final UnitOfMeasurement uom = dataFile.getUnitOfMeasurement(measureValueColumn, line);
+        UnitOfMeasurement uom = dataFile.getUnitOfMeasurement(measureValueColumn, line);
         LOG.debug("UomCode: '{}'", uom);
         // OBSERVED_PROPERTY
-        final ObservedProperty observedProperty = dataFile.getObservedProperty(measureValueColumn, line);
+        ObservedProperty observedProperty = dataFile.getObservedProperty(measureValueColumn, line);
         LOG.debug("ObservedProperty: {}", observedProperty);
         // OFFERING
-        final Offering offer = dataFile.getOffering(sensor);
+        Offering offer = dataFile.getOffering(sensor);
         LOG.debug("Offering: {}", offer);
         // OM:PARAMETER
-        final Optional<List<OmParameter<?>>> omParameter = dataFile.getOmParameters(measureValueColumn, line);
+        Optional<List<OmParameter<?>>> omParameter = dataFile.getOmParameters(measureValueColumn, line);
         return new InsertObservation(sensor,
                 foi,
                 value,
-                timestamp,
+                timeStamp,
                 uom,
                 observedProperty,
                 offer,
