+++ feeder/src/main/java/org/n52/sos/importer/feeder/collector/csv/NSAMParser.java
@@ -26,7 +26,7 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
  * Public License for more details.
  */
-package org.n52.sos.importer.feeder.csv;
+package org.n52.sos.importer.feeder.collector.csv;
 
 import java.io.BufferedReader;
 import java.io.IOException;
